pyproject.toml
[tool.poetry]
name = "flight-search"
version = "0.1.0"
description = "Search flights"
authors = ["zernab hussan <zernabhussain@yahoo.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.2"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
boto3 = "^1.35.43"
jinja2 = "^3.1.4"
python-dotenv = "^1.0.1"
aws-cdk-lib = "2.162.1"
constructs = ">=10.0.0,<11.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "python -m app.main"

[tool.pytest.ini_options]
pythonpath = ["."]